move throttle and stringify to utils/
