+++ src/utils/throttle.ts
@@ -1,0 +1,31 @@
+/**
+ * Throttle a function to run at most once every `wait` milliseconds.
+ */
+export function throttle(fn: () => void, wait: number): () => void {
+  let inCooldown = false
+  let pending = false
+
+  function invoke() {
+    fn()
+    pending = false
+    inCooldown = true
+    // check if there are pending calls after cooldown
+    setTimeout(() => {
+      inCooldown = false
+      if (pending) {
+        // trailing call
+        invoke()
+      }
+    }, wait)
+  }
+
+  return () => {
+    if (!inCooldown) {
+      // leading call
+      invoke()
+    } else {
+      // schedule trailing call
+      pending = true
+    }
+  }
+}
